CM-128: 🔨 디자인 변경에 따른 구조 및 상단 마크업 수정
