+++ styles/global.css
@@ -1,7 +1,6 @@
 * {
   font-family: "NanumSquare", sans-serif;
   box-sizing: border-box;
-  width: 100%;
   padding: 0;
   margin: 0;
